replace numpy.eye in namespace with builtin delta
Namespace expressions have builtin support for identity matrices of arbitrary
size. The unsizedness poses problems in expressions where the lengths of axes
cannot be inferred from (other parts of) the expression, which is the case for
the plane stress elasticity tensor in `plasticity.py`. This problem can be
resolved by defining a fallback length for expressions, which will be used to
define lengths of axes that otherwise would have been undefined. This commit
replaces the `numpy.eye` with the builtin identity matrix `δ`, making use of
the already defined fallback length.