use nutils.function.Custom
In the Nutils v7 development branch the `evaluable` module has been split off
of the `function`. The average user only needs to work with the `function`
module, but in the plasticity script a custom implementation of a
`function.Array` is defined, which, in v7, also requires a corresponding
`evaluable.Array` with `evalf` and `_derivative`. Recently the
`nutils.function.Custom` has been added to the v7 development branch and
backported to v6, which eliminates the need for implementing an
`evaluable.Array`. Instead, defining `evalf` and `partial_derivative` on the
implementation of `function.Custom` is sufficient. This commit replaces the
`function.Array`, `evaluable.Array` pair with one `function.Custom`
implementation.